+++ tests/random_lax_test.py
@@ -67,13 +67,6 @@ class RandomTestBase(jtu.JaxTestCase):
     # kstest does not understand bfloat16 input, so cast to float32.
     if samples.dtype == jnp.bfloat16:
       samples = samples.astype('float32')
-    # kstest fails for infinities starting in scipy 1.12
-    # (https://github.com/scipy/scipy/issues/20386)
-    scipy_version = jtu.parse_version(scipy.__version__)
-    if scipy_version < (1, 14) and np.issubdtype(samples.dtype, np.floating):
-      samples = np.array(samples, copy=True)
-      samples[np.isposinf(samples)] = 0.01 * np.finfo(samples.dtype).max
-      samples[np.isneginf(samples)] = 0.01 * np.finfo(samples.dtype).min
     self.assertGreater(scipy.stats.kstest(samples, cdf).pvalue, fail_prob)
 
   def _CheckChiSquared(self, samples, pmf, *, pval=None):